Revert "Change Render::set_pixel x type to u8, y to u16"
This reverts commit 119be58ae42cba67b99a66b1ed67b90c8e86d837.

Because `RecordRender` need handle canvas width > 255

+++ nes_core/src/nes/ppu.rs
@@ -375,7 +375,8 @@ impl<R: Render> Ppu<R> {
             let pixel = self
                 .effective_mask
                 .apply_effects(self.color_theme.color(pixel_idx));
-            self.renderer.set_pixel(x, self.timing.scanline, pixel.0);
+            self.renderer
+                .set_pixel(x as u32, self.timing.scanline as u32, pixel.0);
         }
 
         // --- vram_addr management (rendering-enabled only) ---
